+++ burst-kotlin-plugin/src/main/kotlin/app/cash/burst/kotlin/FunctionSpecializer.kt
@@ -209,6 +209,9 @@ internal class FunctionSpecializer(
         }
 
     result.annotations += original.testAnnotation.asAnnotation()
+    original.ignoreAnnotation?.let { annotationSymbol ->
+      result.annotations += annotationSymbol.asAnnotation()
+    }
 
     result.irFunctionBody(context = pluginContext) {
       val receiverLocal =